+++ HISTORY.rst
@@ -2,6 +2,18 @@
 History
 =======
 
+
+0.1.5 (2023-XX-XX)
+-------------------------
+* Complete overhaul of ReadTheDocs documentation
+
+0.1.5-alpha (2023-08-15)
+-------------------------
+* Begin overhaul of ReadTheDocs documentation
+* Add notes and example to Koch et al. 2006 metric
+* Rename variable returned by K06 to jet_events_ws
+
+
 0.1.4 (2023-08-02)
 -------------------------
 * Add new metric to package: Zappa et al. 2018 (This method builds on Ceppi et al. 2018)